refactor: post 안에 있는 written 클래스 외부로 꺼냄
