+++ app/src/main/java/com/stopsmoke/kekkek/domain/model/Written.kt
@@ -1,0 +1,8 @@
+package com.stopsmoke.kekkek.domain.model
+
+data class Written(
+    val uid: String,
+    val name: String,
+    val profileImage: ProfileImage,
+    val ranking: Long,
+)
